Working OpenSSL 4 client
